move : /main에 사용하는 컴포넌트 Product폴더로 이동
